ai_service/requirements.txt
google-search-results==2.4.2  # SerpAPI
requests==2.31.0
beautifulsoup4==4.12.3
langchain==0.1.12
langchain-community==0.0.27
python-dotenv==1.0.1
mysql-connector-python==8.3.0
schedule==1.2.1
newspaper3k==0.2.8
tqdm==4.66.2 